laravel add delete function

+++ resources/views/layouts/back/partials/_action.blade.php
@@ -1,3 +1,3 @@
 <a href="{{ $show_url }}" class="btn btn-info btn-xs"><i class="glyphicon glyphicon-eye-open"></i>Show</a>
-<a href="{{ $edit_url }}" class="btn btn-primary btn-xs show-modal edit" title="Edit {{ $model->subject }}"><i class="glyphicon glyphicon-edit"></i>Edit</a>
-<a href="{{ $delete_url }}" class="btn btn-danger btn-xs"><i class="glyphicon glyphicon-trash"></i>Delete</a>
+<a href="{{ $edit_url }}" class="btn btn-primary btn-xs show-modal edit" title="Edit Data"><i class="glyphicon glyphicon-edit"></i>Edit</a>
+<a href="{{ $delete_url }}" class="btn btn-danger btn-xs show-modal-confirm"><i class="glyphicon glyphicon-trash" title="Delete Data"></i>Delete</a>
